% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkLSAY}
\alias{checkLSAY}
\title{checkLSAY}
\usage{
checkLSAY(dataset, cohort)
}
\arguments{
\item{dataset}{The dataset that would be analysed.}

\item{cohort}{year of the cohort (For example, LSAY 2009, cohort = 2009).}
}
\value{
True if the dataset does not have modifications that affect the linkage process.
}
\description{
Check that the dataset is an LSAY dataset
and the dataset does not have modifications affecting the linkage process.
}
