% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkNamesDuplicates}
\alias{checkNamesDuplicates}
\title{checkNamesDuplicates}
\usage{
checkNamesDuplicates(dataset)
}
\arguments{
\item{dataset}{The dataset that would be analysed.}
}
\value{
True if none of the variable names are duplicates, otherwise false if overlap exists.
}
\description{
It's important to check for duplicate names to avoid
errors during data linkage and ensure accurate results.
Stata does not allow duplicate variable names, so this
process ensures the joined datasets don't have any variables with the same name.
}
