% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPostcodeStructure}
\alias{checkPostcodeStructure}
\title{checkPostcodeStructure}
\usage{
checkPostcodeStructure(dataset)
}
\arguments{
\item{dataset}{The dataset that would be analysed.}
}
\value{
True if all the postcodes are valid.
}
\description{
Checks that the postcodes are valid values.
}
