% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkVariableNames}
\alias{checkVariableNames}
\title{checkVariableNames}
\usage{
checkVariableNames(dataset)
}
\arguments{
\item{dataset}{The dataset that would be analysed}
}
\value{
if the variable names are valid. Prints a message describing problem and specific variable that is the problem if invalid.
}
\description{
Check that the variable name is accepted by Stata: Stata variable names must adhere to the following rules:
• Contain 1-32 characters.
• Only contain the characters A-Z, 0-9, and underscore (_).
• Begin with a letter or an underscore.
}
