% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateFolders}
\alias{CreateFolders}
\title{CreateFolders}
\usage{
CreateFolders(path = getwd())
}
\arguments{
\item{name}{the path}
}
\value{
True if the folders are created correctly.
}
\description{
This function generates the folders the script needs to run.
}
