% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{downloadDataverseData}
\alias{downloadDataverseData}
\title{downloadDataverseData}
\usage{
downloadDataverseData(id, file)
}
\arguments{
\item{id}{unique doi (get from dataverse listing of desired data)}

\item{file}{where the file would be storage}
}
\value{
True if the download is correct
}
\description{
Function takes inputs doi (get from dataverse listing of desired data) and the year of the data
}
