% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{FilterConcordance}
\alias{FilterConcordance}
\title{FilterConcordance}
\usage{
FilterConcordance(concordances, id, ratio_threshold = NULL)
}
\arguments{
\item{concordances}{Concorcondaces file.}

\item{id}{Geospatial identificator in que concordances table.}

\item{ratio_threshold}{ratio.}
}
\value{
True if the file is storage correctly.
}
\description{
Narrow down the concordances and select the highest
ratio that meets the threshold set by the user.
}
