% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{GenerateLog}
\alias{GenerateLog}
\title{GenerateLog}
\usage{
GenerateLog(name)
}
\arguments{
\item{name}{the name of the log file}
}
\value{
True if the log file is generated
}
\description{
This function generates the log file that would store all the events generated by the code.
}
