% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabularies.R
\name{GetTerm}
\alias{GetTerm}
\title{GetTerm}
\usage{
GetTerm(term)
}
\arguments{
\item{term}{A String with the direction of the term in the ARDC server.}
}
\value{
Return label, dcterms_created, dcterms_modified, creator, dc_publisher, dc_source, dcterms_title, has_top_concept, history_note, scope_note, type and identifier.
}
\description{
Based on a term, get the metadata associated to it. For example: GetTerm(keywords='http://example.com/census/IFAGEP').
}
