% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{LoadLSAY}
\alias{LoadLSAY}
\title{LoadLSAY}
\usage{
LoadLSAY(cohort, wave, LSAY_topics)
}
\arguments{
\item{cohort}{year of the cohort (For example, LSAY 2009, cohort = 2009)}

\item{wave}{vector of years. (For example, wave = [2012,2013,2014])}

\item{LSAY_topics}{vector with the sub-topics that would be included in the analysis, (For example["School transition" "Current",..]).}
}
\value{
a list which contains the survey data and the geospatial data.
}
\description{
Load or download the LSAY data, given a cohort and wave. (The demonstrator only supports LSAY cohort 2009)
}
