% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{LoadParameters}
\alias{LoadParameters}
\title{LoadParameters}
\usage{
LoadParameters(file)
}
\arguments{
\item{file}{is the path where the JSON file is located.}
}
\value{
True if the parameters are valid
}
\description{
Loads the parameters and sets the global environment.
}
