% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{LoadTSP2021}
\alias{LoadTSP2021}
\title{LoadTSP2021}
\usage{
LoadTSP2021(year = NULL, variables = NULL)
}
\arguments{
\item{year}{Vector of years.}
}
\value{
a list which contains the data and metadata
}
\description{
Load and filter the TSP 2021 by the demonstrator variables and the closest year published census.
For example, if the year is [2017,2018,2019], the TSP2021 only will have the closet census to this years that is 2016.
}
