% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAY_metadata.R
\docType{data}
\name{LSAY_metadata}
\alias{LSAY_metadata}
\title{Metadata - Longitudinal Surveys of Australian Youth}
\format{
## `LSAY_metadata`
A data frame with 37,895 rows and 15 columns:
\describe{
  \item{Cohort}{Y95,Y98,Y03,Y06,Y09,Y15}
  \item{Wave}{}
  \item{Wave/year}{}
  \item{Section}{}
  \item{Data element ID}{}
  \item{Major topic area}{}
  \item{Sub-major topic area}{}
  \item{Minor topic area}{}
  \item{Data element}{}
  \item{Variable}{}
  \item{Type}{}
  \item{Label}{}
  \item{Question}{}
  \item{Base}{}
}
}
\source{
<https://www.lsay.edu.au/publications/search-for-lsay-publications/2621>
}
\usage{
LSAY_metadata
}
\description{
A complete listing of the variables and their associated formats and value labels contained in the LSAY data files for all six LSAY cohorts.
}
\keyword{datasets}
