% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ParsingParameter}
\alias{ParsingParameter}
\title{ParsingParameter}
\usage{
ParsingParameter(parameter)
}
\arguments{
\item{parameter}{String with the name of the parameter.}
}
\value{
The value of the parameter.
}
\description{
Returns the parameters from the global environment.
}
\keyword{internal}
