% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{PotentialCensus}
\alias{PotentialCensus}
\title{PotentialCensus}
\usage{
PotentialCensus(year)
}
\arguments{
\item{year}{year of analysis.}
}
\value{
String which contains the ABS_year and the original year.
}
\description{
Given a year, find the closest census year.
}
\keyword{internal.}
