% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{QualityIndicator}
\alias{QualityIndicator}
\title{QualityIndicator}
\usage{
QualityIndicator(ratio)
}
\arguments{
\item{ratio}{This field describes the Ratio of the FROM region that is being donated to the TO region. The Ratio is a figure between 0 and 1. (ABS, 2021).}
}
\value{
Returns the equivalent SA3 for each row.
}
\description{
Implementation of the ABS quality indicator which provides a considered measure of the quality of the correspondence.
in relation to the weighting unit. (ABS, 2021).
}
