% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabularies.R
\name{SearchConcept}
\alias{SearchConcept}
\title{SearchConcept}
\usage{
SearchConcept(keywords)
}
\arguments{
\item{keywords}{A vector of strings that contains keywords.}
}
\value{
Terms related to the keywords.
}
\description{
Based on a keyword, query the ARDC vocabs server and return related terms. For example: SearchConcept(keywords='AGED').
}
