% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{SpatiallyLSAY}
\alias{SpatiallyLSAY}
\title{SpatiallyLSAY}
\usage{
SpatiallyLSAY(survey, cohort, wave = NULL)
}
\arguments{
\item{survey}{Dataset}

\item{cohort}{year of the cohort (For example, LSAY 2009, cohort = 2009).}

\item{wave}{vector of years. (For example, wave = [2012,2013,2014])}
}
\value{
A dataframe with stata labels intact.
}
\description{
Identify the geospatial table, and filter by the wave and cohorts.
}
\keyword{internal}
