% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{SummaryConcordances}
\alias{SummaryConcordances}
\title{SummaryConcordances}
\usage{
SummaryConcordances(concordances_POA, concordances_SA3)
}
\arguments{
\item{concordances_POA}{concordances using in the POA linkage}

\item{concordances_SA3}{concordances using in the SA3 linkage}
}
\value{
a data frame with the concordances involved in the data linkage.
}
\description{
Create a summary of the concordances involved in the data linkage.
}
