% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SummaryLog}
\alias{SummaryLog}
\title{SummaryLog}
\usage{
SummaryLog(summaryResults)
}
\arguments{
\item{summaryResults}{List which contains the output after the data linkage.}
}
\description{
Consolidate a log with all the information on the data linkage.
}
