% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{SummaryMetrics}
\alias{SummaryMetrics}
\title{SummaryMetrics}
\usage{
SummaryMetrics(metrics_POA, metrics_SA3)
}
\arguments{
\item{metrics_POA}{Metrics created in the POA linkage.}

\item{metrics_SA3}{Metrics created in the SA3 linkage.}
}
\value{
a data frame with all the information mentioned before.
}
\description{
Create a summary with important information about the data linkage, such as year, cohort,
linkage, missing values, not linked areas, not linked individuals and the quality of the data linkage.
}
