% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SummaryReport}
\alias{SummaryReport}
\title{SummaryReport}
\usage{
SummaryReport(concordances_POA, concordances_SA3, metrics_POA, metrics_SA3)
}
\arguments{
\item{concordances_POA}{Concordances used during the first stage: POA to SA3.}

\item{concordances_SA3}{Concordances used during the second stage: SA3 to SA3.}

\item{metrics_POA}{Metrics used during the first stage: POA to SA3.}

\item{metrics_SA3}{Metrics used during the second stage: SA3 to SA3.}
}
\value{
Returns a list with all the information about the data linkage.
}
\description{
Create the summary report
}
