% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{SurveyLSAY}
\alias{SurveyLSAY}
\title{SurveyLSAY}
\usage{
SurveyLSAY(survey, wave = NULL, LSAY_topics = NULL, cohort)
}
\arguments{
\item{wave}{vector of years. (For example, wave = [2012,2013,2014]).}

\item{cohort}{year of the cohort (For example, LSAY 2009, cohort = 2009).}
}
\value{
A dataframe with stata labels intact.
}
\description{
Identify the survey responses, and filter them by the wave and cohorts.
}
\keyword{internal}
