% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{TestDataverseConnection}
\alias{TestDataverseConnection}
\title{TestDataverseConnection}
\usage{
TestDataverseConnection()
}
\value{
True if the connection is correct
}
\description{
Tests connection to ADA dataverse. Requires dataverse token to be loaded in the system environment.
}
