% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{TransformPOA}
\alias{TransformPOA}
\title{TransformPOA}
\usage{
TransformPOA(data, concordances, year, ratio_threshold = NULL)
}
\arguments{
\item{year}{Year.}
}
\value{
Returns the equivalent SA3 for each row.

the metric of missing values, miss matching and POAS that doenst match.
}
\description{
Recieve a vector of POAS, This function transformate POA to SA3.
}
