% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{TransformSA3}
\alias{TransformSA3}
\title{TransformSA3}
\usage{
TransformSA3(data, concordances, year_in, year_out, ratio_threshold)
}
\arguments{
\item{data}{Year.}

\item{concordances}{Year.}
}
\value{
Returns the equivalent SA3 for each row.

the metric of missing values, miss matching and POAS that doenst match.
}
\description{
Recieve a vector of SA3 ABS year,  This function transformate POA to SA3.
}
